feat: [resty] Resort e2e test structure
