+++ test/e2e/metrics/main.go
@@ -1,14 +1,16 @@
-package resty
+package metrics
 
 import (
 	"github.com/gin-gonic/gin"
 	"github.com/prometheus/client_golang/prometheus/promhttp"
-	"math/rand"
-	"time"
+
+	metrics "github.com/quwan-sre/observability-go-contrib/metrics/gin"
 )
 
 func main() {
 	r := gin.Default()
+
+	r.Use(metrics.NewMetricsMiddleware())
 
 	r.GET("/metrics", func(ctx *gin.Context) {
 		promhttp.Handler().ServeHTTP(ctx.Writer, ctx.Request)
@@ -16,7 +18,13 @@ func main() {
 	})
 
 	r.GET("/health", func(ctx *gin.Context) {
-		time.Sleep(time.Duration(rand.Intn(100)) * time.Millisecond)
+		ctx.JSON(200, gin.H{
+			"message": "pong",
+		})
+		return
+	})
+
+	r.GET("/exist", func(ctx *gin.Context) {
 		ctx.JSON(200, gin.H{
 			"message": "pong",
 		})
