+++ test/e2e/metrics/resty_test.go
@@ -1,4 +1,4 @@
-package resty
+package metrics
 
 import (
 	"fmt"
@@ -7,26 +7,7 @@ import (
 	"io"
 	"net/http"
 	"testing"
-	"time"
 )
-
-func TestMain(t *testing.M) {
-	// prepare a gin HTTP server
-	go main()
-
-	// health check
-	for {
-		resp, err := http.Get("http://127.0.0.1:8080/health")
-		if err != nil || resp.StatusCode != 200 {
-			time.Sleep(1 * time.Second)
-		} else {
-			break
-		}
-	}
-
-	// ready, run the test case
-	t.Run()
-}
 
 func TestRestyClient(t *testing.T) {
 	restyClient := resty.New()
